# docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      token:
        type: string
      username:
        type: string
    type: object
...
  handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handlers.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@expense-tracker.local
    name: API Support
    url: http://www.swagger.io/support
  description: Expense Tracker REST API for users, categories, transactions, balances,
    wallets.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Expense Tracker API
  version: "1.0"
paths:
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request password reset token by email
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Forgot password
      tags:
      - Auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Login user
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user and return JWT token
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using reset token
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Reset password
      tags:
      - Auth
  /api/balance:
    get:
      description: Get overall balance for user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get balance
      tags:
      - Balances
  /api/balance/category:
    get:
      description: Get category-wise balance summary for user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get balance by category
      tags:
      - Balances
  /api/balance/monthly:
    get:
      description: Get monthly balance breakdown for user (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get monthly balance
      tags:
      - Balances
  /api/balance/range:
    get:
      description: Get balance summary for date range (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Get balance by date range
      tags:
      - Balances
  /api/balance/recalculate:
    post:
      description: Recalculate user balance from transactions (protected)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Recalculate balance
      tags:
      - Balances
...
schemes:
- http
swagger: "2.0"